+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/podcast/adapter/BookmarkViewHolder.kt
@@ -6,7 +6,10 @@ import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.ComposeView
 import androidx.recyclerview.widget.RecyclerView
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
-import au.com.shiftyjelly.pocketcasts.compose.bookmark.BookmarkItem
+import au.com.shiftyjelly.pocketcasts.compose.bookmark.BookmarkRow
+import au.com.shiftyjelly.pocketcasts.compose.bookmark.BookmarkRowColors
+import au.com.shiftyjelly.pocketcasts.compose.buttons.TimePlayButtonColors
+import au.com.shiftyjelly.pocketcasts.compose.buttons.TimePlayButtonStyle
 import au.com.shiftyjelly.pocketcasts.podcasts.view.podcast.PodcastAdapter
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 
@@ -15,10 +18,10 @@ class BookmarkViewHolder(
     private val theme: Theme,
 ) : RecyclerView.ViewHolder(composeView) {
 
-    fun bind(data: PodcastAdapter.BookmarkItemData,) {
+    fun bind(data: PodcastAdapter.BookmarkItemData) {
         composeView.setContent {
             AppTheme(theme.activeTheme) {
-                BookmarkItem(
+                BookmarkRow(
                     bookmark = data.bookmark,
                     isMultiSelecting = data.isMultiSelecting,
                     isSelected = data.isSelected,
@@ -29,7 +32,11 @@ class BookmarkViewHolder(
                                 onLongPress = { data.onBookmarkRowLongPress(data.bookmark) },
                                 onTap = { data.onBookmarkRowClick(data.bookmark, bindingAdapterPosition) }
                             )
-                        }
+                        },
+                    colors = BookmarkRowColors.Default,
+                    timePlayButtonStyle = TimePlayButtonStyle.Outlined,
+                    timePlayButtonColors = TimePlayButtonColors.Default,
+                    showIcon = true,
                 )
             }
         }