+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/whatsnew/WhatsNewFragment.kt
@@ -49,8 +49,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.compose.AppTheme
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
+import au.com.shiftyjelly.pocketcasts.compose.CallOnce
 import au.com.shiftyjelly.pocketcasts.compose.buttons.RowButton
 import au.com.shiftyjelly.pocketcasts.compose.buttons.RowTextButton
 import au.com.shiftyjelly.pocketcasts.compose.components.TextH20
@@ -64,11 +67,14 @@ import au.com.shiftyjelly.pocketcasts.ui.helper.FragmentHostListener
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.views.fragments.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 import au.com.shiftyjelly.pocketcasts.images.R as IR
 import au.com.shiftyjelly.pocketcasts.localization.R as LR
 
 @AndroidEntryPoint
 class WhatsNewFragment : BaseFragment() {
+
+    @Inject lateinit var analyticsTracker: AnalyticsTrackerWrapper
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -79,6 +85,14 @@ class WhatsNewFragment : BaseFragment() {
             setBackgroundColor(Color.Transparent.toArgb())
             setContent {
                 AppTheme(theme.activeTheme) {
+
+                    CallOnce {
+                        analyticsTracker.track(
+                            AnalyticsEvent.WHATSNEW_SHOWN,
+                            mapOf("version" to Settings.WHATS_NEW_VERSION_CODE)
+                        )
+                    }
+
                     setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
                     val onClose: () -> Unit = {
                         @Suppress("DEPRECATION")
@@ -86,10 +100,20 @@ class WhatsNewFragment : BaseFragment() {
                     }
                     WhatsNewPage(
                         onGoToSettings = {
+                            analyticsTracker.track(
+                                AnalyticsEvent.WHATSNEW_CONFIRM_BUTTON_TAPPED,
+                                mapOf("version" to Settings.WHATS_NEW_VERSION_CODE)
+                            )
                             onClose()
                             goToPlaybackSettings()
                         },
-                        onClose = onClose,
+                        onClose = {
+                            analyticsTracker.track(
+                                AnalyticsEvent.WHATSNEW_DISMISSED,
+                                mapOf("version" to Settings.WHATS_NEW_VERSION_CODE)
+                            )
+                            onClose()
+                        },
                     )
                 }
             }